Реакторинг классов сцен
